split up main.py
