+++ ipypublish/convert/__init__.py
@@ -1,0 +1,3 @@
+#!/usr/bin/env python
+
+# from ipypublish.export_plugins import latex_ipypublish_main
